requirements/base.txt
# Django
Django==1.11.3


# Image
easy-thumbnails==2.4.2
Pillow==4.3.0
# olefile==0.44


# Django Rest Framework
djangorestframework==3.6.3
djoser==0.6.0


# Celery
celery==4.1.0
# amqp==2.1.1
# billiard==3.5.0.3
# kombu==4.1.0
# pytz==2017.2
# vine==1.1.4


# Documentation
mkdocs==0.16.3
# Jinja2==2.9.6
# Markdown==2.6.9
# MarkupSafe==1.0
# PyYAML==3.12
# click==6.7
# livereload==2.5.1
# tornado==4.5.1
